requirements.txt
pysimplegui
PyPDF2
Pillow